1.) reduce some garbage-objects 2.) use Math.toRadians() instead of manual conversion 3.) replace manual copy and fill with System.arraycopy() and Arrays.fill 4.) replace some type var[] by type[] var
git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/batik/trunk@501495 13f79535-47bb-0310-9956-ffa450edef68

+++ sources/org/apache/batik/ext/awt/image/renderable/GaussianBlurRable8Bit.java
@@ -109,7 +109,7 @@ public class GaussianBlurRable8Bit
     /**
      * Constant: 3*sqrt(2*PI)/4
      */
-    static final float DSQRT2PI = (float)(Math.sqrt(2*Math.PI)*3.0/4.0);
+    static final double DSQRT2PI = (Math.sqrt(2*Math.PI)*3.0/4.0);
 
     /**
      * Grow the source's bounds
